+++ plugins/woocommerce-admin/packages/components/src/calendar/date-picker.js
@@ -7,7 +7,6 @@ import { __, sprintf } from '@wordpress/i18n';
 import { createElement, Component } from '@wordpress/element';
 import { Dropdown, DatePicker as WpDatePicker } from '@wordpress/components';
 import { partial } from 'lodash';
-import { TAB } from '@wordpress/keycodes';
 import moment from 'moment';
 import PropTypes from 'prop-types';
 import { dateValidationMessages, toMoment } from '@woocommerce/date';
@@ -26,14 +25,28 @@ class DatePicker extends Component {
 		this.onInputChange = this.onInputChange.bind( this );
 	}
 
-	handleKeyDown( isOpen, onToggle, { keyCode } ) {
-		if ( TAB === keyCode && isOpen ) {
+	handleFocus( isOpen, onToggle ) {
+		if ( ! isOpen ) {
 			onToggle();
 		}
 	}
 
-	handleFocus( isOpen, onToggle ) {
+	handleBlur( isOpen, onToggle, event ) {
 		if ( ! isOpen ) {
+			return;
+		}
+
+		const relatedTargetParent = event.relatedTarget?.closest(
+			'.components-dropdown'
+		);
+		const currentTargetParent = event.currentTarget?.closest(
+			'.components-dropdown'
+		);
+
+		if (
+			! relatedTargetParent ||
+			relatedTargetParent !== currentTargetParent
+		) {
 			onToggle();
 		}
 	}
@@ -81,6 +94,7 @@ class DatePicker extends Component {
 						disabled={ disabled }
 						value={ text }
 						onChange={ this.onInputChange }
+						onBlur={ partial( this.handleBlur, isOpen, onToggle ) }
 						dateFormat={ dateFormat }
 						label={ __( 'Choose a date', 'woocommerce-admin' ) }
 						error={ error }
@@ -98,11 +112,6 @@ class DatePicker extends Component {
 						) }
 						aria-expanded={ isOpen }
 						focusOnMount={ false }
-						onKeyDown={ partial(
-							this.handleKeyDown,
-							isOpen,
-							onToggle
-						) }
 						errorPosition="top center"
 					/>
 				) }